requirements.txt
numpy
pandas
matplotlib.pyplot
seaborn
scikit-learn.model_selection
scikit-learn.tree
scikit-learn.ensemble
scikit-learn.metrics
scikit-learn.linear_model
joblib
termcolor